+++ models/dashboard_model.php
@@ -9,23 +9,17 @@ class Dashboard_Model extends Model
     public function xhrInsert()
     {  
        $text = $_POST['text'];//this will need to be sanitized
-
-       $sth = $this->db->prepare('INSERT INTO data (text) VALUES (:text)');
-       $sth->execute(array(':text' => $text));
-
+        $this->db->insert('data', array('text'=>$text));
         $data = array('text' => $text, 'id' => $this->db->lastInsertId());
-       echo json_encode($data); 
+       echo json_encode($data);
+       $text =''; 
     }
     public function xhrGetListings(){
-        $sth = $this->db->prepare('SELECT * FROM data');
-        $sth->setFetchMode(PDO::FETCH_ASSOC);
-        $sth->execute();
-        $data = $sth->fetchAll();
-        echo json_encode($data);
+        $result = $this->db->select( "SELECT * FROM data");
+        echo json_encode($result);
     }
     public function xhrDeleteListing(){
-        $id = $_POST['id']; //reads the post request
-        $sth = $this->db->prepare('DELETE FROM data WHERE id = "'.$id.'"  ');//uses the ID in the post request to prepare a delete query
-        $sth->execute();//executes the prepared statement    
+        $id = (int) $_POST['id']; 
+        $this->db->delete('data', "id = '$id'");
     }
 }